Move jsx-replace into its own module
